Move read_derivation_path() function from UI module to utils module
