+++ src/signMessage.c
@@ -1,4 +1,3 @@
-#include "getPubkey.h"
 #include "os.h"
 #include "ux.h"
 #include "cx.h"
