+++ src/utils.c
@@ -4,8 +4,32 @@
 #include <stdlib.h>
 #include "utils.h"
 #include "menu.h"
+#include "globals.h"
 
-void get_public_key(uint8_t *publicKeyArray, const uint32_t *derivationPath, size_t pathLength) {
+size_t read_derivation_path(const uint8_t *dataBuffer, size_t size, uint32_t *derivationPath) {
+    if (size == 0) {
+        THROW(ApduReplySolanaInvalidMessage);
+    }
+    size_t len = dataBuffer[0];
+    dataBuffer += 1;
+    if (len < 0x01 || len > BIP32_PATH) {
+        THROW(ApduReplySolanaInvalidMessage);
+    }
+    if (1 + 4 * len > size) {
+        THROW(ApduReplySolanaInvalidMessage);
+    }
+
+    for (unsigned int i = 0; i < len; i++) {
+        derivationPath[i] = ((dataBuffer[0] << 24u) | (dataBuffer[1] << 16u) |
+                             (dataBuffer[2] << 8u) | (dataBuffer[3]));
+        dataBuffer += 4;
+    }
+    return len;
+}
+
+void get_public_key(uint8_t publicKeyArray[PUBKEY_LENGTH],
+                    const uint32_t *derivationPath,
+                    size_t pathLength) {
     cx_ecfp_private_key_t privateKey;
     cx_ecfp_public_key_t publicKey;
